+++ sb3_contrib/common/maskable/evaluation.py
@@ -88,18 +88,21 @@ def evaluate_policy(  # noqa: C901
     current_lengths = np.zeros(n_envs, dtype="int")
     observations = env.reset()
     states = None
-
+    episode_starts = np.ones((env.num_envs,), dtype=bool)
     while (episode_counts < episode_count_targets).any():
         if use_masking:
             action_masks = get_action_masks(env)
             actions, state = model.predict(
                 observations,
                 state=states,
+                episode_start=episode_starts,
                 deterministic=deterministic,
                 action_masks=action_masks,
             )
         else:
-            actions, states = model.predict(observations, state=states, deterministic=deterministic)
+            actions, states = model.predict(
+                observations, state=states, episode_start=episode_starts, deterministic=deterministic
+            )
         observations, rewards, dones, infos = env.step(actions)
         current_rewards += rewards
         current_lengths += 1
@@ -110,6 +113,7 @@ def evaluate_policy(  # noqa: C901
                 reward = rewards[i]
                 done = dones[i]
                 info = infos[i]
+                episode_starts[i] = done
 
                 if callback is not None:
                     callback(locals(), globals())
@@ -133,8 +137,6 @@ def evaluate_policy(  # noqa: C901
                         episode_counts[i] += 1
                     current_rewards[i] = 0
                     current_lengths[i] = 0
-                    if states is not None:
-                        states[i] *= 0
 
         if render:
             env.render()